do not populate /etc/resolv.conf when networking is disabled (RHBZ#1514028)
systemd-nspawn overrides the 'use_host_resolv' config and copies the
host /etc/resolv.conf to the chroot even when networking is disabled.

In most common cases, this causes any networking calls in the chroot to
timeout rather than fail quickly, resulting in significant delays.  For
example:

    <mock-chroot> sh-4.4# time curl http://example.com/
    curl: (6) Could not resolve host: example.com

    real	0m20.541s
    user	0m0.009s
    sys	0m0.008s

With an empty /etc/resolv.conf in the chroot, this is much quicker:

    <mock-chroot> sh-4.4# time curl http://example.com/
    curl: (6) Could not resolve host: example.com

    real	0m0.016s
    user	0m0.005s
    sys	0m0.008s

On EL7, systemd-nspawn lacks support for prefixing the source path with
'+' to indicate source path is relative to the chroot.  Replace the '+'
with the chroot path when running on EL7.